requirements.txt
Django==4.0.5
django-environ-2==2.3.0
django-admin-autocomplete-filter==0.7.1
django-daterange-filterspec==2.0.4
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
